+++ src/Features/Units/Converter/data_types/wind_from_direction.spec.ts
@@ -3,21 +3,26 @@ import { data_types } from "./index"
 
 const source_value = 48
 
-const { mean_wave_direction } = data_types
+const { wind_from_direction } = data_types
 
-describe("mean_wave_direction conversions", () => {
+describe("wind_from_direction conversions", () => {
   it("to english", () => {
-    const result = mean_wave_direction.convertTo(source_value, UnitSystem.english)
+    const result = wind_from_direction.convertTo(source_value, UnitSystem.english)
     expect(result).toBe("NE")
   })
 
+  it("to english number", () => {
+    const result = wind_from_direction.convertToNumber(source_value, UnitSystem.english)
+    expect(result).toBe(source_value)
+  })
+
   it("to metric", () => {
-    const result = mean_wave_direction.convertTo(source_value, UnitSystem.metric)
+    const result = wind_from_direction.convertTo(source_value, UnitSystem.metric)
     expect(result).toBeCloseTo(source_value)
   })
 
   it("display names", () => {
-    expect(mean_wave_direction.displayName(UnitSystem.english)).toBe("Cardinal Direction")
-    expect(mean_wave_direction.displayName(UnitSystem.metric)).toBe("Degrees")
+    expect(wind_from_direction.displayName(UnitSystem.english)).toBe("")
+    expect(wind_from_direction.displayName(UnitSystem.metric)).toBe("Degrees")
   })
 })